+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
 #### Collectors
 - added `intelmq.bots.parsers.opendxl.collector` (#1265).
 - added `intelmq.bots.collectors.api`: collecting data using an HTTP API (#123, #1187).
+- added `intelmq.bots.collectors.rsync` (#1286).
 
 #### Parsers
 - added `intelmq.bots.parsers.mcafee.parser_atd` (#1265).
@@ -33,6 +34,7 @@ CHANGELOG
 ### Packaging
 
 ### Tests
+- Travis: Change the ownership of `/opt/intelmq` to the current user.
 
 ### Tools